+++ gb.qt/src/CTextArea.cpp
@@ -40,10 +40,10 @@ DECLARE_EVENT(EVENT_Change);
 DECLARE_EVENT(EVENT_Cursor);
 DECLARE_EVENT(EVENT_Link);
 
-static void to_pos(QTextEdit *wid, int par, int car, long *pos)
+static void to_pos(QTextEdit *wid, int par, int car, int *pos)
 {
   int i, l;
-  long p = 0;
+  int p = 0;
 
   for (i = 0; i < par; i++)
   {
@@ -57,7 +57,7 @@ static void to_pos(QTextEdit *wid, int par, int car, long *pos)
 }
 
 
-static void from_pos(QTextEdit *wid, long pos, int *par, int *car)
+static void from_pos(QTextEdit *wid, int pos, int *par, int *car)
 {
   int i;
   int l;
@@ -91,10 +91,10 @@ static void look_pos(QTextEdit *wid, int *line, int *col)
 }
 
 
-static void get_selection(QTextEdit *wid, long *start, long *length)
+static void get_selection(QTextEdit *wid, int *start, int *length)
 {
   int pStart, iStart, pEnd, iEnd;
-  long posEnd;
+  int posEnd;
 
   wid->getSelection(&pStart, &iStart, &pEnd, &iEnd);
 
@@ -166,7 +166,7 @@ BEGIN_PROPERTY(CTEXTAREA_alignment)
   if (READ_PROPERTY)
     GB.ReturnInteger(WIDGET->alignment());
   else
-    WIDGET->setAlignment(PROPERTY(long));
+    WIDGET->setAlignment(PROPERTY(int));
 
 END_PROPERTY
 */
@@ -216,7 +216,7 @@ END_PROPERTY
 /*
 BEGIN_PROPERTY(CTEXTAREA_max_length)
 
-  long max;
+  int max;
 
   if (READ_PROPERTY)
   {
@@ -225,7 +225,7 @@ BEGIN_PROPERTY(CTEXTAREA_max_length)
   }
   else
   {
-    max = PROPERTY(long);
+    max = PROPERTY(int);
     if (max <= 0)
       max = -1;
     WIDGET->setMaxLength(max);
@@ -275,7 +275,7 @@ END_PROPERTY
 BEGIN_PROPERTY(CTEXTAREA_pos)
 
   int par, car;
-  long pos;
+  int pos;
 
   if (READ_PROPERTY)
   {
@@ -293,7 +293,7 @@ END_PROPERTY
 
 
 /*
-BEGIN_METHOD(CTEXTAREA_select, long line; long col; long selline; long selcol)
+BEGIN_METHOD(CTEXTAREA_select, int line; int col; int selline; int selcol)
 
   MyMultiLineEdit *wid = QMULTILINEEDIT(_object);
 
@@ -326,9 +326,9 @@ BEGIN_METHOD(CTEXTAREA_insert, GB_STRING text)
 END_METHOD
 
 /*
-BEGIN_METHOD(CTEXTAREA_line_get, long line)
+BEGIN_METHOD(CTEXTAREA_line_get, int line)
 
-  long line = PARAM(line);
+  int line = PARAM(line);
 
   if (line < 0 || line >= WIDGET->numLines())
     GB.ReturnNull();
@@ -364,7 +364,7 @@ END_PROPERTY
 
 BEGIN_PROPERTY(CTEXTAREA_sel_length)
 
-  long start, length;
+  int start, length;
 
   get_selection(WIDGET, &start, &length);
   GB.ReturnInteger(length);
@@ -374,7 +374,7 @@ END_PROPERTY
 
 BEGIN_PROPERTY(CTEXTAREA_sel_start)
 
-  long start, length;
+  int start, length;
 
   get_selection(WIDGET, &start, &length);
   GB.ReturnInteger(start);
@@ -423,7 +423,7 @@ END_METHOD
 
 BEGIN_METHOD(CTEXTAREA_to_pos, GB_INTEGER line; GB_INTEGER col)
 
-  long pos;
+  int pos;
 
   to_pos(WIDGET, VARG(line), VARG(col), &pos);
 
@@ -491,7 +491,7 @@ END_METHOD
 
 /*BEGIN_PROPERTY(CTEXTAREA_scrollbar)
 
-  long scroll;
+  int scroll;
 
   if (READ_PROPERTY)
   {
