+++ main/gbx/gbx_stream_direct.c
@@ -93,7 +93,7 @@ static int stream_close(STREAM *stream)
 }
 
 
-static int stream_read(STREAM *stream, char *buffer, long len)
+static int stream_read(STREAM *stream, char *buffer, int len)
 {
 	return STREAM_read_direct(FD, buffer, len);
 }
@@ -104,19 +104,19 @@ static int stream_getchar(STREAM *stream, char *buffer)
 }
 
 
-static int stream_write(STREAM *stream, char *buffer, long len)
+static int stream_write(STREAM *stream, char *buffer, int len)
 {
 	return STREAM_write_direct(FD, buffer, len);
 }
 
 
-static int stream_seek(STREAM *stream, long long pos, int whence)
+static int stream_seek(STREAM *stream, int64_t pos, int whence)
 {
   return (lseek(FD, pos, whence) < 0);
 }
 
 
-static int stream_tell(STREAM *stream, long long *pos)
+static int stream_tell(STREAM *stream, int64_t *pos)
 {
   *pos = lseek(FD, 0, SEEK_CUR);
   return (*pos < 0);
@@ -141,7 +141,7 @@ static int stream_flush(STREAM *stream)
 #define stream_eof NULL
 
 
-static int stream_lof(STREAM *stream, long long *len)
+static int stream_lof(STREAM *stream, int64_t *len)
 {
   *len = stream->direct.size;
   return FALSE;
